+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## master
 
 * Playing asciicast from stdin: `cat demo.json | asciinema play -`
-* Playing asciicast from IPFS: `asciinema play ipfs://QmcdXYJp6e4zNuimuGeWPwNMHQdxuqWmKx7NhZofQ1nw2V`
+* Playing asciicast from IPFS: `asciinema play ipfs://ipfs/QmcdXYJp6e4zNuimuGeWPwNMHQdxuqWmKx7NhZofQ1nw2V`
 * `-q/--quiet` option added to `rec` command
 
 ## 1.1.1 (2015-06-21)